# api.yml
openapi: 3.0.1
info:
    version: 1.0.0
    title: User Service
    license:
        name: MIT
    contact:
        name: Ronaldo Tantra
        email: ronaldotantra@gmail.com
servers:
    - url: http://localhost
paths:
    /v1/user:
        get:
            summary: Get user
            description: Get user data
            operationId: GetCurrentUser
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ResponseWithData'
                '403':
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '404':
                    description: Not Found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
        patch:
            summary: Update user
            description: Update user data
            operationId: UpdateProfile
            security:
                - bearerAuth: []
            requestBody:
                description: Payload to update user
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PayloadUpdateUser'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BaseResponse'
                '403':
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '409':
                    description: Conflict
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
    /v1/users:
        post:
            summary: Register user
            description: Register new user
            operationId: Register
            requestBody:
                description: Payload to update user
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PayloadInsertUser'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ResponseWithData'
                '409':
                    description: Conflict
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
    /v1/users/login:
        post:
            summary: Login user
            description: Login user
            operationId: Login
            requestBody:
                description: Payload to update user
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PayloadLogin'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ResponseWithData'
                '400':
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'

components:
    schemas:
        PayloadUpdateUser:
            type: object
            required:
                - phone
                - name
            properties:
                phone:
                    type: string
                name:
                    type: string
        PayloadInsertUser:
            type: object
            required:
                - phone
                - name
                - password
            properties:
                phone:
                    type: string
                name:
                    type: string
                password:
                    type: string
        PayloadLogin:
            type: object
            required:
                - phone
                - password
            properties:
                phone:
                    type: string
                password:
                    type: string
        ErrorResponse:
            type: object
            properties:
                code:
                    type: string
                data:
                    type: object
                message:
                    type: string
        BaseResponse:
            type: object
            properties:
                message:
                    type: string
        ResponseWithData:
            type: object
            properties:
                data:
                    type: object
                message:
                    type: string

    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT

x-original-swagger-version: '2.0'